+++ RetroFrog V4/app/components/BuyButton.tsx
@@ -1,25 +1,36 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
+import classNames from 'classnames';
 
 export default function BuyButton() {
-  const [clicked, setClicked] = useState(false);  
-  const [loading, setLoading] = useState(false);  
-  const [purchased, setPurchased] = useState(false);  
+  const [loading, setLoading] = useState(false);
+  const [purchased, setPurchased] = useState(false);
 
   const handleClick = () => {
-    if (clicked || purchased) return; 
-    setClicked(true);  
-    setLoading(true);  
+    if (loading || purchased) return;
+
+    setLoading(true);
+
     setTimeout(() => {
-      setLoading(false);  
-      setPurchased(true);  
-    }, 2000); 
+      setLoading(false);
+      setPurchased(true);
+    }, 2000);
   };
+
+  const isDisabled = loading || purchased;
 
   return (
     <button
       onClick={handleClick}
-      className={`buy-button ${clicked ? 'clicked' : ''} ${loading ? 'loading' : ''} ${purchased ? 'purchased' : ''}`}
-      disabled={clicked || purchased}  
+      className={classNames(
+        'w-full text-textDark font-bold py-2 px-4 rounded-md transition duration-300',
+        'bg-amber-500 hover:bg-amber-600', //Se queda el color verde de base
+        {
+          'buy-button': true,
+          loading: loading,
+          purchased: purchased,
+        },
+      )}
+      disabled={isDisabled}
     >
       {purchased ? 'Ya Adquirido' : loading ? 'Procesando...' : 'Comprar'}
     </button>
