+++ RetroFrog V4/app/components/Custom.tsx
@@ -1,21 +1,17 @@
-import React, { useState } from "react";
+import React, { useState } from 'react';
 
-interface CustomProps {
-  style?: React.CSSProperties;
-  className?: string;
-}
-
-export default function Custom({ style, className }:CustomProps) {
-  
-  const [theme, setTheme] = useState<string>("dark");
-  const [background, setBackground] = useState<string>("/assets/background/bg1.png"); 
+export default function Custom() {
+  const [theme, setTheme] = useState<string>('dark');
+  const [background, setBackground] = useState<string>(
+    '/assets/background/bg.jpg',
+  );
   const [font, setFont] = useState<string>("'Open Sans', sans-serif");
   const [profilePic, setProfilePic] = useState<string | null>(null);
 
-  
   const handleThemeChange = (theme: string) => setTheme(theme);
   const handleFontChange = (e: React.ChangeEvent<HTMLSelectElement>) =>
     setFont(e.target.value);
+
   const handleProfilePicChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     if (e.target.files) {
       const file = e.target.files[0];
@@ -30,51 +26,56 @@ export default function Custom({ style, className }:CustomProps) {
   };
 
   const backgrounds = [
-    "/assets/background/bg1.png",
-    "/assets/background/bg2.png",
-    "/assets/background/bg3.png",
+    '/assets/background/bg1.jpg',
+    '/assets/background/bg2.jpg',
+    '/assets/background/bg3.jpg',
   ];
 
   return (
-    <div
-      className={`h-3/4 w-fit max-w-md bg-[#151A2D] rounded-xl shadow-lg p-4 flex flex-col items-center fixed right-2 top-4 gap-4 transition-all ${className}`}
-      style={{ ...style }}
-    >
-      <h3 className="text-white">Customización de Página</h3>
+    <div className="bg-primaryDark text-textDark flex flex-col w-full h-12">
+      <h3>Customización de Página</h3>
       <div className="flex items-center gap-2">
-      <label htmlFor="theme" className="text-white">Selecciona Tema</label>
-      <select
-        id="theme"
-        value={theme}
-        onChange={(e) => handleThemeChange(e.target.value)}
-        className="hidden"  
-      >
-        <option value="light">Claro</option>
-        <option value="dark">Oscuro</option>
-        <option value="other">Otros</option>
-      </select>
-      <div className="flex gap-2">
-        <button
-          className={`px-3 py-1 rounded ${theme === "light" ? "bg-blue-500" : "bg-gray-500"}`}
-          onClick={() => handleThemeChange("light")}
+        <label htmlFor="theme" className="text-white">
+          Selecciona Tema
+        </label>
+        <select
+          id="theme"
+          value={theme}
+          onChange={(e) => handleThemeChange(e.target.value)}
+          className="hidden"
         >
-          Claro
-        </button>
-        <button
-          className={`px-3 py-1 rounded ${theme === "dark" ? "bg-blue-500" : "bg-gray-500"}`}
-          onClick={() => handleThemeChange("dark")}
-        >
-          Oscuro
-        </button>
-        <button
-          className={`px-3 py-1 rounded ${theme === "other" ? "bg-blue-500" : "bg-gray-500"}`}
-          onClick={() => handleThemeChange("other")}
-        >
-          Otros
-        </button>
+          <option value="light">Claro</option>
+          <option value="dark">Oscuro</option>
+          <option value="other">Otros</option>
+        </select>
+        <div className="flex gap-2">
+          <button
+            className={`px-3 py-1 rounded ${
+              theme === 'light' ? 'bg-blue-500' : 'bg-gray-500'
+            }`}
+            onClick={() => handleThemeChange('light')}
+          >
+            Claro
+          </button>
+          <button
+            className={`px-3 py-1 rounded ${
+              theme === 'dark' ? 'bg-blue-500' : 'bg-gray-500'
+            }`}
+            onClick={() => handleThemeChange('dark')}
+          >
+            Oscuro
+          </button>
+          <button
+            className={`px-3 py-1 rounded ${
+              theme === 'other' ? 'bg-blue-500' : 'bg-gray-500'
+            }`}
+            onClick={() => handleThemeChange('other')}
+          >
+            Otros
+          </button>
+        </div>
       </div>
-    </div>
-      {theme === "other" && (
+      {theme === 'other' && (
         <div className="flex flex-col items-center gap-2">
           <h4 className="text-white">Selecciona un color para el fondo</h4>
           <div className="flex gap-2">
@@ -82,7 +83,10 @@ export default function Custom({ style, className }:CustomProps) {
               <button
                 key={index}
                 className="w-16 h-16 rounded-full"
-                style={{ backgroundImage: `url(${bg})`, backgroundSize: 'cover' }}
+                style={{
+                  backgroundImage: `url(${bg})`,
+                  backgroundSize: 'cover',
+                }}
                 onClick={() => setBackground(bg)}
               ></button>
             ))}
@@ -90,7 +94,9 @@ export default function Custom({ style, className }:CustomProps) {
         </div>
       )}
       <div className="flex items-center gap-2">
-        <label htmlFor="font-family" className="text-white">Fuente</label>
+        <label htmlFor="font-family" className="text-white">
+          Fuente
+        </label>
         <select
           id="font-family"
           value={font}
@@ -104,7 +110,9 @@ export default function Custom({ style, className }:CustomProps) {
         </select>
       </div>
       <div className="flex flex-col items-center gap-2">
-        <label htmlFor="profile-pic" className="text-white">Foto de Perfil</label>
+        <label htmlFor="profile-pic" className="text-white">
+          Foto de Perfil
+        </label>
         <input
           type="file"
           id="profile-pic"
@@ -120,19 +128,6 @@ export default function Custom({ style, className }:CustomProps) {
           />
         )}
       </div>
-      <div
-        className={`w-full h-full p-4 flex flex-col justify-center items-center transition-all rounded-md`}
-        style={{
-          fontFamily: font,
-          backgroundImage: `url(${background})`,
-          backgroundSize: 'cover',
-          color: theme === "dark" ? "white" : "black",
-        }}
-      >
-        <p className="text-center">
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aenean nec arcu augue. Aliquam erat volutpat. Suspendisse vitae lobortis mi. Curabitur semper suscipit lectus sed tempor. Integer vulputate dolor at sodales egestas. In bibendum, odio fringilla congue eleifend, nulla sem dictum magna, non convallis lacus ligula id magna. Phasellus velit dui, gravida eget nunc non, consectetur ultricies lacus. Duis malesuada mauris nec tellus volutpat, sit amet sollicitudin orci fringilla. Donec accumsan purus in mi finibus ornare. Aenean in condimentum lorem. Fusce consequat luctus metus id posuere. Nam vitae fringilla enim.
-        </p>
-      </div>
     </div>
   );
-};
+}
