+++ RetroFrog V4/app/tailwind.css
@@ -12,7 +12,7 @@ body {
 }
 
 body {
-  background-image: url('../assets/background/bg.jpg');
+  background-image: url('/assets/background/bg.jpg');
   background-size: cover;
   background-attachment: fixed;
   background-position: center;
